refactor into new project directory structure
