+++ dinghy_ping/views/templates/base.html
