+++ dinghy_ping/views/templates/index.html
@@ -26,7 +26,7 @@
                         </label>
                         <p class="text-black text-xs italic">Provide optional headers as JSON key/value format.</p>
                         <br>
-                        <img class="logo float-right" src="static/silvermullet_small_1.jpg" title="silvermullet" />
+                        <img class="logo float-right" src="../static/silvermullet_small_1.jpg" title="silvermullet" />
                         <input
                             class="cursor-auto bg-green-300 hover:bg-green-600 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline"
                             type="submit" value="Ping">
