+++ dinghy_ping/views/templates/ping_response.html
