+++ dinghy_ping/views/templates/pod_logs.html
@@ -24,7 +24,7 @@
                                     class="shadow appearance-none border border-red rounded w-full py-2 px-3 text-gray-700 mb-3 leading-tight focus:outline-none focus:shadow-outline"
                                     id="tail_lines" type="text" placeholder="100" name="tail_lines">
                             </label>
-                            <img class="logo float-right" src="/static/silvermullet_small_1.jpg" title="silvermullet" />
+                            <img class="logo float-right" src="../static/silvermullet_small_1.jpg" title="silvermullet" />
                             <input
                                 class="cursor-auto bg-green-300 hover:bg-green-600 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline"
                                 type="submit" value="Pods">
