+++ dinghy_ping/views/templates/pod_logs_output.html
